+++ sentry/src/main/java/io/sentry/TracePropagationTargets.java
@@ -8,11 +8,11 @@ import org.jetbrains.annotations.NotNull;
  * Checks if an URL matches the list of origins to which `sentry-trace` header should be sent in
  * HTTP integrations.
  */
-public final class TracingOrigins {
+public final class TracePropagationTargets {
 
   public static boolean contain(final @NotNull List<String> origins, final @NotNull String url) {
     if (origins.isEmpty()) {
-      return true;
+      return false;
     }
     for (final String origin : origins) {
       if (url.contains(origin) || url.matches(origin)) {